Image Functionality!
check_price(): Changes
Added filtering to locate and save the image url of the landing image to a variable. Used requests to download said image to the premade directory.

GUI(): Changes
Added drop down menu which gets its selectioins from a list, currently working on getting the selections to autopopulate the rest of the GUI with the product name, price, image and graph.

Hard coded the first image I've downloaded to figure out how to properly incorporate it into the GUI, had to add another inport to get image handling working but the image is now fully integrated into the gui and is auto-resizing when the program is stretched/maximized.